requirements.txt
python-dotenv
langchain
langchain_community
openai
huggingface_hub
streamlit